Finalne izmene

+++ app/Http/Controllers/ProviderController.php
@@ -128,8 +128,12 @@ class ProviderController extends Controller
     public function destroy(Provider $provider)
     {
         if(auth()->user()->isUser())
-            return response()->json('You are not authorized to delete providers.'); 
+            return response()->json('You are not authorized to delete providers.');
             
+        $apprat = AppointmentRating::get()->where('provider', $provider_id);
+        if (count($apprat) > 0)
+            return response()->json('You cannot delete providers that have appointment ratings.');
+
         $provider->delete();
 
         return response()->json('Provider is deleted successfully.');
